requirements.txt
click==7.1.2
Flask==1.1.2
future==0.18.2
futures==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.15.0
python-whois==0.7.3
Werkzeug==1.0.1
